+++ CHANGELOG.md
@@ -5,11 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Update BDK to version 0.20.0
 
+## [v0.8.0]
+- Update BDK to version 0.20.0 [#169]
 - APIs Added
-  - `TxBuilder.add_data(data: Vec<u8>)`
-  - `Wallet.list_unspent()` returns `Vec<LocalUtxo>`
+  - `TxBuilder.add_data(data: Vec<u8>)` [#163]
+  - `Wallet.list_unspent()` returns `Vec<LocalUtxo>` [#158]
+  - Add coin control methods on TxBuilder [#164]
+
+[#163]: https://github.com/bitcoindevkit/bdk-ffi/pull/163
+[#158]: https://github.com/bitcoindevkit/bdk-ffi/pull/158
+[#164]: https://github.com/bitcoindevkit/bdk-ffi/pull/164
+[#169]: https://github.com/bitcoindevkit/bdk-ffi/pull/169
 
 ## [v0.7.0]
 - Update BDK to version 0.19.0
@@ -58,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.2.0]
 
-[unreleased]: https://github.com/bitcoindevkit/bdk-ffi/compare/v0.7.0...HEAD
+[unreleased]: https://github.com/bitcoindevkit/bdk-ffi/compare/v0.8.0...HEAD
+[v0.8.0]: https://github.com/bitcoindevkit/bdk-ffi/compare/v0.7.0...v0.8.0
 [v0.7.0]: https://github.com/bitcoindevkit/bdk-ffi/compare/v0.6.0...v0.7.0
 [v0.6.0]: https://github.com/bitcoindevkit/bdk-ffi/compare/v0.5.0...v0.6.0
 [v0.5.0]: https://github.com/bitcoindevkit/bdk-ffi/compare/v0.4.0...v0.5.0